// frontend/src/app/api/model/containerReadSectorVersion.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReadSectorVersion } from './readSectorVersion';


export interface ContainerReadSectorVersion { 
    objects?: Array<ReadSectorVersion>;
    totalCount?: number;
}